% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{calculate_dv}
\alias{calculate_dv}
\title{Calculate verification numer of a RUT}
\usage{
calculate_dv(num)
}
\arguments{
\item{num}{RUT to calculate}
}
\value{
Expected verification number
}
\description{
Calculate verification numer of a RUT
}
\examples{
calculate_dv("12345678")
}
\keyword{internal}
