% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clear_rut}
\alias{clear_rut}
\title{Clear RUT}
\usage{
clear_rut(rut)
}
\arguments{
\item{rut}{Character string}
}
\value{
Returns a character string without said characters
}
\description{
Removes any character that is not a number, K, k, dot or hyphen
}
\examples{
clear_rut('12m345.678-9=')
}
\keyword{clean}
