% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rut_complete}
\alias{rut_complete}
\title{RUT format: Dots and Hyphen (Puntos y guion)}
\usage{
rut_complete(rut, uppercase = FALSE)
}
\arguments{
\item{rut}{RUT to modify. It must be a character string}

\item{uppercase}{Defines if the verification number 'k' should be in uppercase (TRUE) or lowercase (FALSE). By default the value is FALSE}
}
\value{
Returns a RUT in dots and hyphen format
}
\description{
Changes the format of a RUT to dots and hyphen
}
\examples{
rut_complete('123456789')
}
\keyword{format}
