% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rut_hyphen}
\alias{rut_hyphen}
\title{RUT format: Hyphen without dots (Sin puntos, con guion)}
\usage{
rut_hyphen(rut, uppercase = FALSE)
}
\arguments{
\item{rut}{RUT to modify. It must be a character string}

\item{uppercase}{Defines if the verification number 'k' should be in uppercase (TRUE) or lowercase (FALSE). By default the value is FALSE}
}
\value{
Returns a RUT in hyphen without dots format
}
\description{
Changes the format of a RUT to hyphen without dots
}
\examples{
rut_hyphen('123456789')
}
\keyword{format}
