% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.R
\name{rut_simple}
\alias{rut_simple}
\title{RUT format: Without dots and Hyphen (Sin puntos ni guion)}
\usage{
rut_simple(rut, uppercase = FALSE)
}
\arguments{
\item{rut}{RUT to modify. It must be a character string}

\item{uppercase}{Defines if the verification number 'k' should be in uppercase (TRUE) or lowercase (FALSE). By default the value is FALSE}
}
\value{
Returns a RUT whitout dots and hyphen
}
\description{
Changes the format of a RUT to not contain dots or hyphen
}
\examples{
rut_simple('123456789')
}
\keyword{format}
