% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate_rut}
\alias{validate_rut}
\title{Validate RUT}
\usage{
validate_rut(rut)
}
\arguments{
\item{rut}{Character str}
}
\value{
Returs TRUE if RUT is valid, else it will return FALSE
}
\description{
Validates a Chilean RUT. The input must be a character string in one of the following formats: RUT with dots and hyphen (Ex. '12.345.678-9'), RUT without dots but with hyphen (Ex. '12345678-9') or RUT without dots and hyphen (Ex. '123456789')
}
\examples{
validate_rut('123456789')
validate_rut('12345678-9')
validate_rut('12.345.678-9')
}
\keyword{validate}
